Publish a version of the style with non-relative URLs
This avoids the need to mangle URLs and makes it easier for
sites like the wiki to add vector tiles.

The old /demo paths are kept working to avoid breaking the
website before it points at the new URLs.

The style file being published is currently only intended for
OSM use so it has CORS headers only set for OSM domains and
127.0.0.1.

The demo pages have been partially moved over to the new
paths but not completely. That will be done after the /styles/
paths are deployed on the OSMF servers.